+++ sv-parser-pp/src/preprocess.rs
@@ -209,7 +209,11 @@ pub fn preprocess_str<T: AsRef<Path>, U: AsRef<Path>>(
             _ => (),
         }
         match n {
-            NodeEvent::Enter(RefNode::ResetallCompilerDirective(_)) if !skip => {}
+            NodeEvent::Enter(RefNode::ResetallCompilerDirective(x)) if !skip => {
+                let locate: Locate = x.try_into().unwrap();
+                let range = Range::new(locate.offset, locate.offset + locate.len);
+                ret.push(locate.str(&s), Some((path.as_ref(), range)));
+            }
             NodeEvent::Enter(RefNode::UndefineCompilerDirective(x)) if !skip => {
                 let (_, _, ref name) = x.nodes;
                 let id = identifier((&name.nodes.0).into(), &s).unwrap();